Refactoring legacy code into legacy.py
