+++ src/pykakasi/kanji.pyi
@@ -1,16 +1,10 @@
+from .properties import Configurations as Configurations
 from typing import Any, Tuple
 
-from .properties import Configurations as Configurations
-from .scripts import H2 as H2
-
-class J2:
-    convert: Any = ...
-    def __init__(self, mode: str = ..., method: str = ...) -> None: ...
+class JConv:
+    def __init__(self) -> None: ...
     def isRegion(self, c: str) -> Any: ...
-    def isCletter(self, literal: str, c: str) -> bool: ...
-    def convert_h(self, itext: Any) -> Tuple[str, int]: ...
-    def convert_nonh(self, text: Any): ...
-    def convert_noop(self, text: Any): ...
+    def convert(self, itext: str) -> Tuple[str, int]: ...
 
 class Itaiji:
     __dict__: Any = ...
